==> code/10-AnalysisDeathsvsEmissions.ipynb <==
# Analysis Deaths-Pollutants

import pandas as pd
import numpy  as np
import csv
import matplotlib.pyplot as plt
import seaborn as sns
from ipywidgets import interact, interactive, fixed, interact_manual
from glob import glob

Para visualizar más Contaminantes poner un símbolo de "#" delante del fichero en el directorio abajo indicado, sólo se leen dos por ser objeto de estudio

#Listado de ficheros entrada DeathsEmissions por emisión
ListFichDeathsEmissionsFinal = glob('../data/csv/final/#DeathsEmissions_final_*.csv')

#Genera fichero de agrupación por contaminantes, causa muerte 1, 2 y 3 y Sexo con totales defunciones 
DfTotalContCausSexo = pd.DataFrame()

for i in range(0,len(ListFichDeathsEmissionsFinal)):   
    
    print ('Contaminante.: '+ListFichDeathsEmissionsFinal[i])
            
    DfContCausSexo = pd.DataFrame()
    
    DfDeathsEmissions = pd.read_csv(ListFichDeathsEmissionsFinal[i],sep=';',compression='gzip') #cargamos df con cada fichero

    DfContCausSexo['Contaminante']    = DfDeathsEmissions['Contaminante']
    DfContCausSexo['CausaMuertebas1'] = DfDeathsEmissions['CausaMuertebas1']
    DfContCausSexo['CausaMuertebas2'] = DfDeathsEmissions['CausaMuertebas2']
    DfContCausSexo['CausaMuertebas3'] = DfDeathsEmissions['CausaMuertebas3']
    DfContCausSexo['Sexo']            = DfDeathsEmissions['Sexo']

    Columnas = ('Contaminante',
                'CausaMuertebas1',
                'CausaMuertebas2',
                'CausaMuertebas3',
                'Sexo')   

    DfContCausSexo = DfContCausSexo.drop_duplicates(Columnas, keep="first", inplace=False)
   
    DfTotales = DfDeathsEmissions.groupby(Columnas).agg({'Id':'count'})
    DfTotales = DfTotales.rename(columns={"Id":"TotalDeaths"})
    DfContCausSexo = DfContCausSexo.merge(DfTotales, how='left', left_on=Columnas, right_index=True)

    DfTotalContCausSexo = DfTotalContCausSexo.append(DfContCausSexo)


#Listado CIE-10
DfCie10 = pd.read_csv('../data/csv/CIE-10.csv',sep=';')

#Incluye descripción causa de muerte y descripción sexo
DfTotalContCausSexo = pd.merge(DfTotalContCausSexo, DfCie10, on='CausaMuertebas1')
DfTotalContCausSexo['SexoDescripcion'] = 'Varón'
DfTotalContCausSexo['SexoDescripcion'][DfTotalContCausSexo['Sexo'] == 6] = 'Mujer'

DfTotalContCausSexo['ContaminanteN'] = pd.Categorical(DfTotalContCausSexo['Contaminante']).codes 

ListaContaminantes = DfTotalContCausSexo['Contaminante'].unique()

def FContaminante(Contaminante):
    Df1 = DfTotalContCausSexo[DfTotalContCausSexo['Contaminante'] == Contaminante]
    sns.barplot(data=Df1, x='Descripcion', y='TotalDeaths', hue='SexoDescripcion')
    sns.color_palette("RdBu", n_colors=5)
    plt.suptitle('Total Defunciones por Contaminante y Causa de Muerte año 2015') 
    plt.xlabel('Causas de Muerte') 
    plt.ylabel('Total Defunciones') 
    plt.xticks(rotation = 90)
    plt.show()
    

interact(FContaminante,Contaminante = ListaContaminantes)
